requirements.txt
amqp==5.3.1
anyascii==0.3.3
asgiref==3.11.0
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
defusedxml==0.7.1
Django==6.0.1
django-ckeditor-5==0.2.19
django-cors-headers==4.9.0
django-filter==25.2
django-modelcluster==6.4.1
django-permissionedforms==0.1
django-stubs-ext==5.2.9
django-taggit==6.1.0
django-tasks==0.9.0
django-treebeard==4.8.0
djangorestframework==3.16.1
draftjs_exporter==5.2.0
et_xmlfile==2.0.0
filetype==1.2.0
GDAL==3.8.4
gunicorn==25.0.0
idna==3.11
kombu==5.6.2
laces==0.1.2
modelsearch==1.1.1
openpyxl==3.1.5
packaging==26.0
pillow==12.1.0
pillow_heif==1.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
redis==7.1.0
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
telepath==0.3.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wagtail==7.3
wcwidth==0.5.3
Willow==1.12.0
